refactor: clean up some code duplications
