+++ apps/web/src/pages/meine-mathe-skills/index.tsx
@@ -53,11 +53,11 @@ function Content() {
               <p>5. Klasse</p>
             </a>
             <a
-              href="/meine-mathe-skills/grundwissen-realschule-bayern"
+              href="/meine-mathe-skills/training-realschule-bayern"
               onClick={(e) => {
                 e.preventDefault()
                 void router.push(
-                  '/meine-mathe-skills/grundwissen-realschule-bayern',
+                  '/meine-mathe-skills/training-realschule-bayern',
                   undefined,
                   {
                     shallow: true,
@@ -73,9 +73,11 @@ function Content() {
               <p>
                 Training
                 <br />
-                Mittlere Reife
+                Realschule
                 <br />
-                Bayern (RS)
+                Bayern
+                <br />
+                (Abschlussprüfung)
               </p>
             </a>
           </div>
